requirements.txt
fastapi==0.109.2
pydantic==2.6.1
pydantic-settings==2.1.0
uvicorn==0.27.1
python-multipart==0.0.7
PyMuPDF==1.23.26
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
loguru==0.7.2 